Adding Test Environment and Player prefab
test.environment.js is complete, and a basic player prefab has been made. Feel free to play around in it and add things if needed. (Also as a side note, it has bgm, but Chrome has problems with autoplaying it. You might have to click the canvas to start playing it)